XFormInstanceManager.get_forms
