+++ corehq/form_processor/backends/sql/dbaccessors.py
@@ -35,6 +35,7 @@ from corehq.form_processor.exceptions import (
     XFormNotFound,
     XFormSaveError,
 )
+from corehq.form_processor.models.util import sort_with_id_list as _sort_with_id_list
 from corehq.form_processor.interfaces.dbaccessors import (
     AbstractCaseAccessor,
     AbstractLedgerAccessor,
@@ -382,14 +383,8 @@ class FormAccessorSQL:
 
     @staticmethod
     def get_forms(form_ids, ordered=False):
-        assert isinstance(form_ids, list)
-        if not form_ids:
-            return []
-        forms = list(XFormInstance.objects.plproxy_raw('SELECT * from get_forms_by_id(%s)', [form_ids]))
-        if ordered:
-            _sort_with_id_list(forms, form_ids, 'form_id')
-
-        return forms
+        """DEPRECATED"""
+        return XFormInstance.objects.get_forms(form_ids, ordered)
 
     @staticmethod
     def get_attachments(form_id):
@@ -1483,22 +1478,6 @@ class LedgerAccessorSQL(AbstractLedgerAccessor):
             )
             for trans in resultset:
                 yield trans
-
-
-def _sort_with_id_list(object_list, id_list, id_property):
-    """Sort object list in the same order as given list of ids
-
-    SQL does not necessarily return the rows in any particular order so
-    we need to order them ourselves.
-
-    NOTE: this does not return the sorted list. It sorts `object_list`
-    in place using Python's built-in `list.sort`.
-    """
-    def key(obj):
-        return index_map[getattr(obj, id_property)]
-
-    index_map = {id_: index for index, id_ in enumerate(id_list)}
-    object_list.sort(key=key)
 
 
 def _attach_prefetch_models(objects_by_id, prefetched_models, link_field_name, cached_attrib_name):
